# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CBSE Class 9 Education App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Subject {
  id: string;
  name: string;
  description: string;
  icon: string;
  color: string;
  chapterCount: number;
}

export interface Chapter {
  id: string;
  subjectId: string;
  number: number;
  title: string;
  description: string;
  youtubeVideoId: string;
  duration: string;
}

export interface ChapterDetail {
  id: string;
  subjectId: string;
  number: number;
  title: string;
  description: string;
  youtubeVideoId: string;
  duration: string;
  subjectName: string;
}

export interface ChapterSummary {
  chapterId: string;
  summary: string;
  keyPoints: string[];
}

export interface McqQuestion {
  id: string;
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}