# lib/api-zod/src/generated/types/chapterDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CBSE Class 9 Education App API
 * OpenAPI spec version: 0.1.0
 */

export interface ChapterDetail {
  id: string;
  subjectId: string;
  number: number;
  title: string;
  description: string;
  youtubeVideoId: string;
  duration: string;
  subjectName: string;
}