# lib/api-zod/src/generated/types/mcqQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CBSE Class 9 Education App API
 * OpenAPI spec version: 0.1.0
 */

export interface McqQuestion {
  id: string;
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}