# lib/api-zod/src/generated/types/subject.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CBSE Class 9 Education App API
 * OpenAPI spec version: 0.1.0
 */

export interface Subject {
  id: string;
  name: string;
  description: string;
  icon: string;
  color: string;
  chapterCount: number;
}